Converted test_arithmetic_expression to use GoogleTest
